[Redux] Move account states from queriesSlice to accountsSlice
