+++ packages/desktop-client/src/accounts/accountsSlice.ts
@@ -8,12 +8,12 @@ import {
   type TransactionEntity,
   type SyncServerSimpleFinAccount,
   type SyncServerPluggyAiAccount,
+  type CategoryEntity,
 } from 'loot-core/types/models';
 
 import { resetApp } from '@desktop-client/app/appSlice';
 import { addNotification } from '@desktop-client/notifications/notificationsSlice';
 import {
-  markAccountsDirty,
   markPayeesDirty,
   setNewTransactions,
 } from '@desktop-client/queries/queriesSlice';
@@ -27,11 +27,21 @@ type AccountState = {
     [key: AccountEntity['id']]: { type: string; code: string };
   };
   accountsSyncing: Array<AccountEntity['id']>;
+  updatedAccounts: Array<AccountEntity['id']>;
+  accounts: AccountEntity[];
+  isAccountsLoading: boolean;
+  isAccountsLoaded: boolean;
+  isAccountsDirty: boolean;
 };
 
 const initialState: AccountState = {
   failedAccounts: {},
   accountsSyncing: [],
+  updatedAccounts: [],
+  accounts: [],
+  isAccountsLoading: false,
+  isAccountsLoaded: false,
+  isAccountsDirty: false,
 };
 
 type SetAccountsSyncingPayload = {
@@ -45,6 +55,14 @@ type MarkAccountFailedPayload = {
 };
 
 type MarkAccountSuccessPayload = {
+  id: AccountEntity['id'];
+};
+
+type MarkUpdatedAccountsPayload = {
+  ids: AccountState['updatedAccounts'];
+};
+
+type MarkAccountReadPayload = {
   id: AccountEntity['id'];
 };
 
@@ -70,11 +88,149 @@ const accountsSlice = createSlice({
     ) {
       delete state.failedAccounts[action.payload.id];
     },
+    markUpdatedAccounts(
+      state,
+      action: PayloadAction<MarkUpdatedAccountsPayload>,
+    ) {
+      state.updatedAccounts = action.payload.ids
+        ? [...state.updatedAccounts, ...action.payload.ids]
+        : state.updatedAccounts;
+    },
+    markAccountRead(state, action: PayloadAction<MarkAccountReadPayload>) {
+      state.updatedAccounts = state.updatedAccounts.filter(
+        id => id !== action.payload.id,
+      );
+    },
+    markAccountsDirty(state) {
+      _markAccountsDirty(state);
+    },
   },
   extraReducers: builder => {
     builder.addCase(resetApp, () => initialState);
+
+    builder.addCase(createAccount.fulfilled, _markAccountsDirty);
+    builder.addCase(updateAccount.fulfilled, _markAccountsDirty);
+    builder.addCase(closeAccount.fulfilled, _markAccountsDirty);
+    builder.addCase(reopenAccount.fulfilled, _markAccountsDirty);
+
+    builder.addCase(reloadAccounts.fulfilled, (state, action) => {
+      _loadAccounts(state, action.payload);
+    });
+
+    builder.addCase(reloadAccounts.rejected, state => {
+      state.isAccountsLoading = false;
+    });
+
+    builder.addCase(reloadAccounts.pending, state => {
+      state.isAccountsLoading = true;
+    });
+
+    builder.addCase(getAccounts.fulfilled, (state, action) => {
+      _loadAccounts(state, action.payload);
+    });
+
+    builder.addCase(getAccounts.rejected, state => {
+      state.isAccountsLoading = false;
+    });
+
+    builder.addCase(getAccounts.pending, state => {
+      state.isAccountsLoading = true;
+    });
   },
 });
+type CreateAccountPayload = {
+  name: string;
+  balance: number;
+  offBudget: boolean;
+};
+
+export const createAccount = createAppAsyncThunk(
+  `${sliceName}/createAccount`,
+  async ({ name, balance, offBudget }: CreateAccountPayload) => {
+    const id = await send('account-create', {
+      name,
+      balance,
+      offBudget,
+    });
+    return id;
+  },
+);
+
+type CloseAccountPayload = {
+  id: AccountEntity['id'];
+  transferAccountId?: AccountEntity['id'];
+  categoryId?: CategoryEntity['id'];
+  forced?: boolean;
+};
+
+export const closeAccount = createAppAsyncThunk(
+  `${sliceName}/closeAccount`,
+  async ({
+    id,
+    transferAccountId,
+    categoryId,
+    forced,
+  }: CloseAccountPayload) => {
+    await send('account-close', {
+      id,
+      transferAccountId: transferAccountId || undefined,
+      categoryId: categoryId || undefined,
+      forced,
+    });
+  },
+);
+
+type ReopenAccountPayload = {
+  id: AccountEntity['id'];
+};
+
+export const reopenAccount = createAppAsyncThunk(
+  `${sliceName}/reopenAccount`,
+  async ({ id }: ReopenAccountPayload) => {
+    await send('account-reopen', { id });
+  },
+);
+
+type UpdateAccountPayload = {
+  account: AccountEntity;
+};
+
+export const updateAccount = createAppAsyncThunk(
+  `${sliceName}/updateAccount`,
+  async ({ account }: UpdateAccountPayload) => {
+    await send('account-update', account);
+    return account;
+  },
+);
+
+export const getAccounts = createAppAsyncThunk(
+  `${sliceName}/getAccounts`,
+  async () => {
+    // TODO: Force cast to AccountEntity.
+    // Server is currently returning the DB model it should return the entity model instead.
+    const accounts = (await send('accounts-get')) as unknown as AccountEntity[];
+    return accounts;
+  },
+  {
+    condition: (_, { getState }) => {
+      const { account } = getState();
+      return (
+        !account.isAccountsLoading &&
+        (account.isAccountsDirty || !account.isAccountsLoaded)
+      );
+    },
+  },
+);
+
+export const reloadAccounts = createAppAsyncThunk(
+  `${sliceName}/reloadAccounts`,
+  async () => {
+    // TODO: Force cast to AccountEntity.
+    // Server is currently returning the DB model it should return the entity model instead.
+    const accounts = (await send('accounts-get')) as unknown as AccountEntity[];
+    return accounts;
+  },
+);
 
 type UnlinkAccountPayload = {
   id: AccountEntity['id'];
@@ -83,10 +239,9 @@ type UnlinkAccountPayload = {
 export const unlinkAccount = createAppAsyncThunk(
   `${sliceName}/unlinkAccount`,
   async ({ id }: UnlinkAccountPayload, { dispatch }) => {
-    const { markAccountSuccess } = accountsSlice.actions;
     await send('account-unlink', { id });
-    dispatch(markAccountSuccess({ id }));
-    dispatch(markAccountsDirty());
+    dispatch(actions.markAccountSuccess({ id }));
+    dispatch(actions.markAccountsDirty());
   },
 );
 
@@ -241,11 +396,11 @@ export const syncAccounts = createAppAsyncThunk(
       return false;
     }
 
-    const queriesState = getState().queries;
+    const { accounts } = getState().account;
     let accountIdsToSync: string[];
     if (id === 'offbudget' || id === 'onbudget') {
       const targetOffbudget = id === 'offbudget' ? 1 : 0;
-      accountIdsToSync = queriesState.accounts
+      accountIdsToSync = accounts
         .filter(
           ({ bank, closed, tombstone, offbudget }) =>
             !!bank && !closed && !tombstone && offbudget === targetOffbudget,
@@ -256,7 +411,7 @@ export const syncAccounts = createAppAsyncThunk(
       accountIdsToSync = [id];
     } else {
       // Default: all accounts
-      accountIdsToSync = queriesState.accounts
+      accountIdsToSync = accounts
         .filter(
           ({ bank, closed, tombstone }) => !!bank && !closed && !tombstone,
         )
@@ -339,9 +494,10 @@ export const syncAccounts = createAppAsyncThunk(
       setNewTransactions({
         newTransactions,
         matchedTransactions,
-        updatedAccounts,
       }),
     );
+
+    dispatch(markUpdatedAccounts({ ids: updatedAccounts }));
 
     // Reset the sync state back to empty (fallback in case something breaks
     // in the logic above)
@@ -367,6 +523,12 @@ export const moveAccount = createAppAsyncThunk(
 export const { name, reducer, getInitialState } = accountsSlice;
 export const actions = {
   ...accountsSlice.actions,
+  createAccount,
+  updateAccount,
+  getAccounts,
+  reloadAccounts,
+  closeAccount,
+  reopenAccount,
   linkAccount,
   linkAccountSimpleFin,
   linkAccountPluggyAi,
@@ -374,3 +536,26 @@ export const actions = {
   unlinkAccount,
   syncAccounts,
 };
+
+export const {
+  markAccountRead,
+  markAccountFailed,
+  markAccountSuccess,
+  markAccountsDirty,
+  markUpdatedAccounts,
+  setAccountsSyncing,
+} = accountsSlice.actions;
+
+function _loadAccounts(
+  state: AccountState,
+  accounts: AccountState['accounts'],
+) {
+  state.accounts = accounts;
+  state.isAccountsLoading = false;
+  state.isAccountsLoaded = true;
+  state.isAccountsDirty = false;
+}
+
+function _markAccountsDirty(state: AccountState) {
+  state.isAccountsDirty = true;
+}
