+++ packages/desktop-client/src/components/sidebar/Account.tsx
@@ -25,6 +25,10 @@ import { type AccountEntity } from 'loot-core/types/models';
 
 import { BalanceHistoryGraph } from './BalanceHistoryGraph';
 
+import {
+  reopenAccount,
+  updateAccount,
+} from '@desktop-client/accounts/accountsSlice';
 import { Link } from '@desktop-client/components/common/Link';
 import { Notes } from '@desktop-client/components/Notes';
 import {
@@ -40,10 +44,6 @@ import { useDragRef } from '@desktop-client/hooks/useDragRef';
 import { useNotes } from '@desktop-client/hooks/useNotes';
 import { useSyncedPref } from '@desktop-client/hooks/useSyncedPref';
 import { openAccountCloseModal } from '@desktop-client/modals/modalsSlice';
-import {
-  reopenAccount,
-  updateAccount,
-} from '@desktop-client/queries/queriesSlice';
 import { useDispatch } from '@desktop-client/redux';
 import { type SheetFields, type Binding } from '@desktop-client/spreadsheet';
