+++ packages/desktop-client/src/global-events.ts
@@ -2,6 +2,7 @@
 import { listen } from 'loot-core/platform/client/fetch';
 import * as undo from 'loot-core/platform/client/undo';
 
+import { reloadAccounts } from './accounts/accountsSlice';
 import { setAppState } from './app/appSlice';
 import { closeBudgetUI } from './budgets/budgetsSlice';
 import { closeModal, pushModal, replaceModal } from './modals/modalsSlice';
@@ -10,11 +11,7 @@ import {
   addNotification,
 } from './notifications/notificationsSlice';
 import { loadPrefs } from './prefs/prefsSlice';
-import {
-  reloadAccounts,
-  reloadCategories,
-  reloadPayees,
-} from './queries/queriesSlice';
+import { reloadCategories, reloadPayees } from './queries/queriesSlice';
 import { type AppStore } from './redux/store';
 import * as syncEvents from './sync-events';
