+++ packages/desktop-client/src/hooks/useAccounts.ts
@@ -2,13 +2,13 @@ import { useEffect } from 'react';
 
 import { useInitialMount } from './useInitialMount';
 
-import { getAccounts } from '@desktop-client/queries/queriesSlice';
+import { getAccounts } from '@desktop-client/accounts/accountsSlice';
 import { useSelector, useDispatch } from '@desktop-client/redux';
 
 export function useAccounts() {
   const dispatch = useDispatch();
   const isInitialMount = useInitialMount();
-  const isAccountsDirty = useSelector(state => state.queries.isAccountsDirty);
+  const isAccountsDirty = useSelector(state => state.account.isAccountsDirty);
 
   useEffect(() => {
     if (isInitialMount || isAccountsDirty) {
@@ -16,5 +16,5 @@ export function useAccounts() {
     }
   }, [dispatch, isInitialMount, isAccountsDirty]);
 
-  return useSelector(state => state.queries.accounts);
+  return useSelector(state => state.account.accounts);
 }
