+++ packages/desktop-client/src/hooks/useUpdatedAccounts.ts
@@ -1,5 +1,5 @@
 import { useSelector } from '@desktop-client/redux';
 
 export function useUpdatedAccounts() {
-  return useSelector(state => state.queries.updatedAccounts);
+  return useSelector(state => state.account.updatedAccounts);
 }
