+++ packages/desktop-client/src/modals/modalsSlice.ts
@@ -566,7 +566,7 @@ export const openAccountCloseModal = createAppAsyncThunk(
         id: accountId,
       },
     );
-    const account = getState().queries.accounts.find(
+    const account = getState().account.accounts.find(
       acct => acct.id === accountId,
     );
